Change insights to explore; Add redirect url for no place/topic case
