+++ server/lib/explore/detector.py
