+++ server/routes/explore/html.py
@@ -23,13 +23,24 @@ from flask import render_template
 bp = Blueprint('explore', __name__, url_prefix='/explore')
 
 
-@bp.route('/<string:topic>')
-def page(topic):
+@bp.route('/')
+def page():
   # No production support yet.
   if os.environ.get('FLASK_ENV') == 'production':
     flask.abort(404)
 
   return render_template('/explore.html',
+                         maps_api_key=current_app.config['MAPS_API_KEY'],
+                         website_hash=os.environ.get("WEBSITE_HASH"))
+
+
+@bp.route('/<string:topic>')
+def landing(topic):
+  # No production support yet.
+  if os.environ.get('FLASK_ENV') == 'production':
+    flask.abort(404)
+
+  return render_template('/explore_landing.html',
                          topic=topic,
                          maps_api_key=current_app.config['MAPS_API_KEY'],
                          website_hash=os.environ.get("WEBSITE_HASH"))
