+++ server/templates/explore_landing.html
@@ -17,25 +17,21 @@
 
  {% set is_hide_full_footer = true %}
  {% set is_hide_sub_footer = true %}
- {% set main_id = 'nl-interface' %}
- {% set page_id = 'nl-interface-page' %}
+ {% set main_id = 'explore-landing' %}
+ {% set page_id = 'explore-landing-page' %}
  {% set title = 'Explore' %}
- {% set subpage_title = '' %}
- {% set subpage_url = url_for('nl.page') %}
 
  {% block head %}
- <link rel="stylesheet" href={{url_for('static', filename='css/insights.min.css', t=config['VERSION'])}} >
+ <link rel="stylesheet" href={{url_for('static', filename='css/explore_landing.min.css', t=config['VERSION'])}} >
  <link rel="stylesheet" href="https://fonts.googleapis.com/icon?family=Material+Icons">
- <link rel="stylesheet" href="https://fonts.googleapis.com/icon?family=Material+Icons+Outlined">
- <script src="https://datacommons.org/datacommons.js"></script>
  {% endblock %}
 
  {% block content %}
- <div id="dc-insights-interface"></div>
+ <div id="dc-explore-landing"></div>
  {% endblock %}
  {% block footer %}
  <script
-   src={{url_for('static', filename='insights.js', t=config['VERSION'])}}></script>
+   src={{url_for('static', filename='explore_landing.js', t=config['VERSION'])}}></script>
  <script src="https://maps.googleapis.com/maps/api/js?key={{maps_api_key}}&libraries=places&language={{locale}}" async
    defer></script>
  {% endblock %}
