--- static/css/insights.scss
@@ -1,151 +1,0 @@
-/**
- * Copyright 2023 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-@use "./shared/subject_page";
-@import "./tools/base_tools";
-@import "./draw";
-@import "./draw_choropleth";
-@import "./draw_scatter";
-@import "./shared/story_block";
-@import "./shared/story_chart";
-
-#main-header {
-  height: 72px;
-}
-
-#insight-lhs {
-  padding-right: 20px;
-}
-
-.insights-container {
-  display: flex;
-  flex-direction: column;
-  h1 {
-    font-size: 2rem;
-    margin: 3rem 0;
-  }
-  .insights-charts {
-    font-size: 1.25rem;
-  }
-}
-
-.search-section {
-  margin: 20px 0;
-  .btn {
-    height: 2.5rem;
-    border-top-left-radius: 0;
-    border-bottom-left-radius: 0;
-  }
-  .history-button {
-    margin-left: 0.5rem;
-  }
-}
-
-.topics-box {
-  margin-top: 1rem;
-  padding-top: 1rem;
-  border-top: 1px solid #efefef;
-
-  .topics-head {
-    font-weight: 500;
-    font-size: 0.9rem;
-    color: var(--dc-gray);
-  }
-
-  .topic-link {
-    color: var(--dc-gray);
-    font-size: 13px;
-    margin-right: 4px;
-    text-decoration: underline;
-    color: var(--dc-gray);
-    display: inline-block;
-  }
-}
-
-#child-places {
-  margin-top: 1rem;
-  padding-top: 1rem;
-  border-top: 1px solid #efefef;
-
-  #child-place-head {
-    font-weight: 500;
-    font-size: 0.9rem;
-    color: var(--dc-gray);
-  }
-
-  .child-place-group {
-    margin-bottom: 1rem;
-    margin-top: 0.5rem;
-  }
-
-  .child-place-type {
-    color: var(--dc-gray);
-    font-weight: 600;
-    font-size: 13px;
-  }
-
-  .child-place-link {
-    color: var(--dc-gray);
-    font-size: 13px;
-    margin-right: 4px;
-    text-decoration: underline;
-    color: var(--dc-gray);
-    display: inline-block;
-  }
-}
-
-#parent-places {
-  #parent-place-head {
-    font-weight: 500;
-    font-size: 0.9rem;
-    color: var(--dc-gray);
-  }
-
-  .parent-place-link {
-    color: var(--dc-gray);
-    font-size: 13px;
-    text-decoration: underline;
-    color: var(--dc-gray);
-    display: inline-block;
-  }
-}
-
-#place-callout {
-  font-weight: bold;
-  font-size: 25px;
-}
-
-.topic-item {
-  width: 100%;
-}
-
-.sub-topic-group {
-  margin-left: 10px;
-}
-
-#user-message {
-  margin-top: 1rem;
-  padding-top: 1rem;
-  width: 100%;
-  font-weight: 500;
-  font-size: 0.9rem;
-  color: var(--dc-gray);
-}
-
-.nav-item.highlight {
-  background-color: rgba(0, 0, 0, 0.05);
-  border-radius: 0.5rem;
-}
