+++ static/js/apps/explore/sidebar.tsx
@@ -41,7 +41,7 @@ export function Sidebar(props: SidebarPropType): JSX.Element {
     <div id="subject-page-sidebar">
       <ul id="nav-topics" className="nav flex-column accordion">
         {props.peerTopics.map((topic, idx) => {
-          const url = `/insights/#t=${topic.dcid}&p=${props.place}&pcmp=${props.cmpPlace}`;
+          const url = `/explore/#t=${topic.dcid}&p=${props.place}&pcmp=${props.cmpPlace}`;
           const isCurrentTopic = topic.dcid == props.currentTopicDcid;
           // Do not display itself for "root".
           const isRootTopic = topic.dcid == "dc/topic/Root";
@@ -70,7 +70,7 @@ export function Sidebar(props: SidebarPropType): JSX.Element {
                     if (!category.dcid || category.dcid === topic.dcid) {
                       return;
                     }
-                    const url = `/insights/#t=${category.dcid}&p=${props.place}&pcmp=${props.cmpPlace}`;
+                    const url = `/explore/#t=${category.dcid}&p=${props.place}&pcmp=${props.cmpPlace}`;
                     return (
                       <a
                         key={randDomId()}
